fix(NLS): enhances error handling routines
When a user attempts to run `Get-NLSSite` on a customer that has not yet been configured,
the service will return 404 (symbolizing no sites were found). PowerShell properly
considers this to be a failed HTTP response and errors accordingly. This commit causes
`Get-NLSSite` to ignore 404 response codes.

Additionally, if a transient error occurs that does not have an expected JSON body, the
current exception handler will fail to parse the error and bail; hiding the exception and
leaving the user with no usable information. This commit will now throw the original
exception if the message is not present or doesn't appear to be JSON on trivial glance.